+++ client/src/components/PetCard.tsx
@@ -3,17 +3,9 @@ import cat from "../assets/cat.png";
 import other from "../assets/otherAnimalsBtn.svg"
 import paw from "../assets/favourite-paw.svg";
 import locationPin from "../assets/location-pin.svg"
-import styles from "./PetCard.module.css";
 
-interface CardItems {
-    name: string; 
-    tags: string;
-    location: string; 
-    breed: string;
-    type: string; 
-    age: number;
-    photo: string;
-}
+import styles from "./PetCard.module.css";
+import { CardItems } from "../types";
 
 export default function PetCard( {name, tags, location, type, age, photo} :CardItems ) {
